I know why it's fucked up now...
There are several "special" segments that you cannot just copy over.
The first load segment includes everything from the beginning of the file, so that needs to be considered.
Also, the PHDR segment cannot be copied over either because it must actually point to the program header table...
So many headers as datas as headers D: