Problem with 32 bit executables.
Was using a hack to add a -1 before .ctors, but sometimes it isn't loaded into memory.
The new way of infecting is to change the program entry point.
This means the payload is responsible for giving control back to the victim.
There is an example of this in samples/hello-payload-64.s, but all other sample payloads are currently broken.